+++ iroh-rpc-client/Cargo.toml
@@ -18,7 +18,6 @@ cid = "0.8.0"
 config = "0.13.1"
 futures = "0.3.21"
 iroh-metrics = { path = "../iroh-metrics", default-features = false }
-iroh-rpc-types = { path = "../iroh-rpc-types", default-features = false }
 iroh-util = { path = "../iroh-util" }
 libp2p = { version = "0.50", default-features = false, features = ["gossipsub"] }
 paste = "1.0.7"
@@ -32,14 +31,13 @@ tonic = { version = "0.8", optional = true }
 tonic-health = { version = "0.7", optional = true }
 tower = { version = "0.4.13", optional = true }
 tracing = "0.1.34"
-
+iroh-rpc-types = { path = "../iroh-rpc-types", optional = true }
 
 [dev-dependencies]
 iroh-rpc-types = { path = "../iroh-rpc-types", features = ["testing"] }
-tokio-stream = { version = "0.1.9", features = ["net"]}
+tokio-stream = { version = "0.1.9", features = ["net"] }
 
 [features]
 default = ["grpc", "mem"]
-grpc = ["tonic", "tonic-health", "iroh-rpc-types/grpc", "iroh-metrics/rpc-grpc", "tower"]
-qrpc = ["quic-rpc"]
+grpc = ["tonic", "tonic-health", "iroh-rpc-types/grpc", "iroh-rpc-types/mem", "iroh-metrics/rpc-grpc", "tower"]
 mem = ["iroh-rpc-types/mem"]
