+++ iroh-rpc-client/src/grpc/mod.rs
@@ -1,0 +1,13 @@
+pub mod client;
+pub mod config;
+pub mod gateway;
+pub mod network;
+#[cfg(feature = "grpc")]
+pub mod status;
+pub mod store;
+pub use client::Client;
+
+pub use self::config::Config;
+pub use network::{Lookup, P2pClient};
+pub use status::{ServiceStatus, StatusRow, StatusTable};
+pub use store::StoreClient;
