+++ iroh-rpc-types/Cargo.toml
@@ -20,7 +20,7 @@ multiaddr = "0.16.0"
 paste = "1.0.7"
 prost = "0.11"
 prost-types = "0.11.1"
-quic-rpc = { path = "../../quic-rpc", optional = true }
+quic-rpc = { path = "../../quic-rpc" }
 serde = { version = "1" }
 serde_with = "2.0.0"
 tokio = { version = "1", features = ["net", "sync"] }
@@ -33,7 +33,7 @@ prost-build = "0.11.1"
 tonic-build = { version = "0.8", features = ["prost"], optional = true }
 
 [features]
-default = ["grpc", "mem"]
+default = []
 # GRPC based transport
 grpc = [
   "tonic",
@@ -44,7 +44,6 @@ grpc = [
 ]
 # Memory based transport
 mem = []
-qrpc = ["quic-rpc"]
 
 # Builds and empty tonic server & client as well as associated protobufs
 # that can be used for testing
