+++ iroh-rpc-types/src/grpc/mod.rs
@@ -1,7 +1,9 @@
 pub mod addr;
 pub mod gateway;
-pub mod store;
 pub mod p2p;
+pub mod store;
 
 #[cfg(feature = "testing")]
 pub mod test;
+
+pub use addr::Addr;
