+++ iroh-rpc-types/src/qrpc/mod.rs
@@ -2,3 +2,5 @@ pub mod addr;
 pub mod gateway;
 pub mod p2p;
 pub mod store;
+
+pub use addr::Addr;
